Move MILA geometry helpers out of rewards
